feat: extract UdonRadioCommunicationEditor to dedicated Editor assembly
- Move UdonRadioCommunicationEditor out of UdonRadioCommunication.cs into
  Scripts/Editor/UdonRadioCommunicationEditor.cs with its own asmdef
  (UdonRadioCommunication.Editor, includePlatforms: Editor) — modern U# pattern
- Replace deprecated FindObjectsOfType<UdonBehaviour>() with
  FindObjectsByType<UdonBehaviour>(FindObjectsSortMode.None) (Unity 2022 API)
- Replace .Select(u => u as T).Where(u => u != null) with .OfType<T>()
- Fix BuildCallback : Editor, IVRCSDKBuildRequestedCallback →
  BuildCallback : IVRCSDKBuildRequestedCallback only (no spurious Editor base class)
- Add explicit "Setup" button in OnInspectorGUI for manual triggering
- No #if guards needed in the new editor file — handled by asmdef platform restriction

Agent-Logs-Url: https://github.com/esnya/UdonRadioCommunications/sessions/b4b26279-c0dc-45c1-9d81-d7fbb20ac463
