+++ Packages/com.nekometer.esnya.udon-radio-communications/Scripts/Editor/UdonRadioCommunication.Editor.asmdef
@@ -1,0 +1,20 @@
+{
+    "name": "UdonRadioCommunication.Editor",
+    "references": [
+        "UdonRadioCommunication.Runtime",
+        "VRC.Udon",
+        "UdonSharp.Runtime",
+        "UdonSharp.Editor"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": false,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
